requirements.txt
discord==2.3.2
discord.py==2.5.2
openpyxl==3.1.5
python-dotenv==1.1.0
